+++ .gitignore
@@ -6,3 +6,4 @@
 cf
 cf32
 cf64
+/cf.vcxproj.user
